requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.1.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
multidict==4.7.6
pytz==2020.1
requests==2.24.0
six==1.15.0
slackclient==2.8.1
slacker==0.14.0
soupsieve==2.0.1
tzlocal==2.1
urllib3==1.25.10
websocket-client==0.44.0
yarl==1.5.1
slackeventsapi==2.2.1
Flask==1.1.2